+++ tests/DependencyInjection/AjgarlagPsrHttpMessageExtensionTest.php
@@ -26,6 +26,7 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
         $config = [];
 
         $extension->load([$config], $container);
+        $extension->process($container);
 
         $this->assertTrue($container->has(HttpMessageFactoryInterface::class));
         $this->assertTrue($container->has(HttpFoundationFactoryInterface::class));
@@ -52,6 +53,7 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
         ];
 
         $extension->load([$config], $container);
+        $extension->process($container);
 
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_message_factory'));
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_foundation_factory'));
@@ -66,6 +68,10 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
 
     public function testAjgarlagPsrHttpMessageAliasConfiguration()
     {
+        if (!class_exists(Psr7ServerRequestResolver::class)) {
+            $this->markTestSkipped('Require sensio/framework-extra-bundle:>=5.3 <6');
+        }
+
         $container = new ContainerBuilder();
         $extension = new AjgarlagPsrHttpMessageExtension();
         $config = [
@@ -76,6 +82,7 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
 
         $container->setParameter('ajgarlag_psr_http_message_sensio_psr7_enabled', true);
         $extension->load([$config], $container);
+        $extension->process($container);
 
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_message_factory'));
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_foundation_factory'));